sec-edgar-main/requirements.txt
pandas==2.3.2
requests==2.31.0
pyppeteer==2.0.0
pytest==8.4.2
pytest-asyncio==1.2.0
python-dotenv==1.1.1